Updated by Lin Chen

+++ src/main/java/mum/edu/service/IAdvisorService.java
@@ -5,10 +5,10 @@ import java.util.List;
 import mum.edu.domain.Advisor;
 
 public interface IAdvisorService {
-	public int addAdvisor(Advisor advisor);
-	public void deleteAdvisor(int id);
-	public void updateAdvisor(int id,Advisor advisor);
-	public Advisor getAdvisor(int id);
+	public long addAdvisor(Advisor advisor);
+	public void deleteAdvisor(long id);
+	public void updateAdvisor(long id,Advisor advisor);
+	public Advisor getAdvisor(long id);
     public List<Advisor> getAllAdvisor();
     public List<TMReportService> getStudentTMRport(int studentid);
     public List<TMReportService> getAllStudentTMReport();
